+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/list/immutable/ImmutableQuintupletonList.java
@@ -61,7 +61,7 @@ final class ImmutableQuintupletonList<T>
     @Override
     public T get(int index)
     {
-        switch (index)
+        switch (((Integer)(index)).intValue())
         {
             case 0:
                 return this.element1;
